fix: change float to int

+++ Assets/1_FeedTheRealm/UI/MenuBar/FileOption/PublishMenu/PublishMenuController.cs
@@ -788,7 +788,7 @@ namespace FeedTheRealm.UI.MenuBar.FileOption.PublishMenu
             string categoryName,
             string fullPath,
             string existingSpriteId,
-            float price
+            int price
         )
         {
             if (!string.IsNullOrEmpty(existingSpriteId))
